requirements.txt
asyncio
python-dotenv
python-telegram-bot
wakeonlan
